+++ lib/gtk.dart
@@ -1,5 +1,4 @@
 library gtk;
 
-export 'src/utils/utils.dart';
 export 'src/models/models.dart';
 export 'src/widgets/widgets.dart';