+++ lib/src/widgets/widgets.dart
@@ -3,6 +3,7 @@ export 'gtk_header_bar.dart';
 export 'gtk_header_button.dart';
 export 'gtk_popup_menu.dart';
 export 'gtk_sidebar.dart';
+export 'gtk_theme.dart';
 export 'gtk_two_pane.dart';
 export 'gtk_view_switcher.dart';
 export 'gtk_view_switcher_tab.dart';